+++ cmd/promviz/app/gauge.go
@@ -58,20 +58,3 @@ func RenderGauge(
 
 	return result
 }
-
-func RenderStat(
-	title string,
-	value string,
-	gridPos dashboard.GridPos,
-	viewport *viewport.Model,
-) string {
-
-	// padding := 2
-	colWidth := viewport.Width / 24
-	totalWidth := colWidth * gridPos.W
-
-	// Combine gauge and value
-	result := lipgloss.NewStyle().Width(totalWidth).Align(lipgloss.Center).Render(value)
-
-	return result
-}
